product와 discount 커스텀훅 분리
